requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
psycopg2_binary==2.9.10
python_dotenv==1.0.1